+++ slap-saver/middlewares/checkCode.js
@@ -5,7 +5,6 @@ const checkCode = async (req, res, next) => {
 	}
 	const currentUser = await getCurrentUser(req.user?.id);
 	if (currentUser.position !== 4) {
-		alert('권한이 없습니다.');
 		res.redirect('/author');
 	}
 	next();
