+++ slap-saver/routes/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 const router = express.Router();
 
-const userCtrl = require('../controllers/userController');
+const userCtrl = require('../controllers/user/userController');
 
 /* GET home page. */
 // NOTE: 기본 홈 화면
