+++ slap-saver/views/user/article.ejs
@@ -25,8 +25,8 @@
     <div></div>
     <div class="article__control__right-buttons">
       <!--TODO:공유,북마크 기능 herf 추가-->
-      <button class="article__control__right-buttons__share-button"></button>
-      <button class="article__control__right-buttons__bookmark-button"></button>
+      <button id="btn_share" class="article__control__right-buttons__share-button"></button>
+      <button id="btn_bookmark" class="article__control__right-buttons__bookmark-button"></button>
     </div>
   </div>
 </div>
